refactor(player): extract pure helpers into slideshow_engine module (#235)
Phase 4 PR-1 of agora#226. Lifts the pure-Python helpers
(_parse_schema_version, _is_forward_schema, _parse_iso8601_utc,
_locate_slide_at) and constants (CLOCK_SKEW_TOLERANCE_S,
RESYNC_CAP_MS, PLAYER_MAX_MANIFEST_SCHEMA_VERSION) out of
player/service.py into a new player/slideshow_engine.py module.

These helpers are intentionally pure-Python with no GLib / Gst /
mpv / Linux-path dependencies, so they can be re-used unchanged
by agora-softplayer (Windows host).

service.py re-imports them as the original underscored names, so
all existing call sites (and the 30+ tests that reach for
svc._parse_schema_version etc.) continue to work with zero
modification. No behavior change.

The SlideshowEngine class extraction that this is paving the way
for lands in PR-2; the softplayer migration that consumes it lands
in PR-3 (agora-softplayer).

117 tests pass locally (test_slideshow_player + test_chromium_backend
+ test_slideshow_fetch).
